Rename ClassificationObject to InputObject
