+++ src/Sql/Sql/DataClassification/Cmdlet/DisableAzSqlInstanceDatabaseSensitivityRecommendation.cs
@@ -20,16 +20,16 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Cmdlet
     [Cmdlet(
         VerbsLifecycle.Disable,
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + DataClassificationCommon.SqlInstanceDatabaseSensitivityRecommendation,
-        DefaultParameterSetName = DataClassificationCommon.ClassificationObjectParameterSet,
+        DefaultParameterSetName = DataClassificationCommon.InputObjectParameterSet,
         SupportsShouldProcess = true),
         OutputType(typeof(bool))]
-    public class DisableAzSqlInstanceDatabaseSensitivityRecommendation : ModifyAzSqlInstanceDatabaseSensitivityClassificationCmdlet
+    public class DisableAzSqlInstanceDatabaseSensitivityRecommendation : ModifyAzSqlInstanceDatabaseSensitivityRecommendationCmdlet
     {
         protected override ManagedDatabaseSensitivityClassificationModel GetEntity()
         {
-            if (ClassificationObject != null)
+            if (InputObject != null)
             {
-                return ClassificationObject;
+                return InputObject;
             }
             else if (DatabaseObject != null)
             {
