+++ src/Sql/Sql/DataClassification/Cmdlet/EnableAzSqlInstanceDatabaseSensitivityRecommendation.cs
@@ -20,16 +20,16 @@ namespace Microsoft.Azure.Commands.Sql.DataClassification.Cmdlet
     [Cmdlet(
         VerbsLifecycle.Enable,
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + DataClassificationCommon.SqlInstanceDatabaseSensitivityRecommendation,
-        DefaultParameterSetName = DataClassificationCommon.ClassificationObjectParameterSet,
+        DefaultParameterSetName = DataClassificationCommon.InputObjectParameterSet,
         SupportsShouldProcess = true),
         OutputType(typeof(bool))]
-    public class EnableAzSqlInstanceDatabaseSensitivityRecommendation : ModifyAzSqlInstanceDatabaseSensitivityClassificationCmdlet
+    public class EnableAzSqlInstanceDatabaseSensitivityRecommendation : ModifyAzSqlInstanceDatabaseSensitivityRecommendationCmdlet
     {
         protected override ManagedDatabaseSensitivityClassificationModel GetEntity()
         {
-            if (ClassificationObject != null)
+            if (InputObject != null)
             {
-                return ClassificationObject;
+                return InputObject;
             }
             else if (DatabaseObject != null)
             {
