+++ src/Sql/Sql/help/Disable-AzSqlInstanceDatabaseSensitivityRecommendation.md
@@ -13,10 +13,10 @@ Disables (dismisses) sensitivity recommendations on columns in the Azure SQL man
 
 ## SYNTAX
 
-### ClassificationObjectParameterSet (Default)
+### InputObjectParameterSet (Default)
 ```
 Disable-AzSqlInstanceDatabaseSensitivityRecommendation
- -ClassificationObject <ManagedDatabaseSensitivityClassificationModel> [-PassThru] [-AsJob]
+ -InputObject <ManagedDatabaseSensitivityClassificationModel> [-PassThru] [-AsJob]
  [-DefaultProfile <IAzureContextContainer>] [-WhatIf] [-Confirm] [<CommonParameters>]
 ```
 
@@ -68,21 +68,6 @@ Required: False
 Position: Named
 Default value: None
 Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -ClassificationObject
-An object representing a SQL Managed Instance Database Sensitivity Classification.
-
-```yaml
-Type: Microsoft.Azure.Commands.Sql.DataClassification.Model.ManagedDatabaseSensitivityClassificationModel
-Parameter Sets: ClassificationObjectParameterSet
-Aliases:
-
-Required: True
-Position: Named
-Default value: None
-Accept pipeline input: True (ByValue)
 Accept wildcard characters: False
 ```
 
@@ -143,6 +128,21 @@ Required: False
 Position: Named
 Default value: None
 Accept pipeline input: False
+Accept wildcard characters: False
+```
+
+### -InputObject
+An object representing a SQL Managed Instance Database Sensitivity Classification.
+
+```yaml
+Type: Microsoft.Azure.Commands.Sql.DataClassification.Model.ManagedDatabaseSensitivityClassificationModel
+Parameter Sets: InputObjectParameterSet
+Aliases:
+
+Required: True
+Position: Named
+Default value: None
+Accept pipeline input: True (ByValue)
 Accept wildcard characters: False
 ```
