+++ src/Sql/Sql/help/Enable-AzSqlDatabaseSensitivityRecommendation.md
@@ -13,9 +13,9 @@ Enables sensitivity recommendations on columns (recommendations are enabled by d
 
 ## SYNTAX
 
-### ClassificationObjectParameterSet (Default)
+### InputObjectParameterSet (Default)
 ```
-Enable-AzSqlDatabaseSensitivityRecommendation -ClassificationObject <SqlDatabaseSensitivityClassificationModel>
+Enable-AzSqlDatabaseSensitivityRecommendation -InputObject <SqlDatabaseSensitivityClassificationModel>
  [-PassThru] [-AsJob] [-DefaultProfile <IAzureContextContainer>] [-WhatIf] [-Confirm] [<CommonParameters>]
 ```
 
@@ -62,21 +62,6 @@ Required: False
 Position: Named
 Default value: None
 Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -ClassificationObject
-An object representing a SQL Database Sensitivity Classification.
-
-```yaml
-Type: Microsoft.Azure.Commands.Sql.DataClassification.Model.SqlDatabaseSensitivityClassificationModel
-Parameter Sets: ClassificationObjectParameterSet
-Aliases:
-
-Required: True
-Position: Named
-Default value: None
-Accept pipeline input: True (ByValue)
 Accept wildcard characters: False
 ```
 
@@ -137,6 +122,21 @@ Required: False
 Position: Named
 Default value: None
 Accept pipeline input: False
+Accept wildcard characters: False
+```
+
+### -InputObject
+An object representing a SQL Database Sensitivity Classification.
+
+```yaml
+Type: Microsoft.Azure.Commands.Sql.DataClassification.Model.SqlDatabaseSensitivityClassificationModel
+Parameter Sets: InputObjectParameterSet
+Aliases:
+
+Required: True
+Position: Named
+Default value: None
+Accept pipeline input: True (ByValue)
 Accept wildcard characters: False
 ```
